improve(Etablissement): renomme les composants et personnalise la sidebar pour suivre les maquettes (#6903)
