+++ 2024-frontend/src/layouts/LayoutSidebarCanteen.vue
@@ -4,6 +4,7 @@ import { computedAsync } from "@vueuse/core"
 import { computed } from "vue"
 import urlService from "@/services/urls.js"
 import canteenService from "@/services/canteens.js"
+import AppSeparator from "@/components/AppSeparator.vue"
 
 /* Route */
 const route = useRoute()
@@ -25,15 +26,15 @@ const canteenBadgeGroupe = computed(() => {
 
 /* Sidebar links */
 const menuItems = computed(() =>  {
-  const informationActive = currentRoute.value === "GestionnaireBilanInformations"
-  const gestionnairesActive = currentRoute.value === "GestionnaireBilanGestionnaires"
-  const pagePubliqueActive = currentRoute.value === "GestionnaireBilanPagePublique"
-  const toutesTeledeclarationsActive = currentRoute.value === "GestionnaireBilanToutesTeledeclarations"
-  const cantinesGroupeActive = currentRoute.value === "GestionnaireBilanCantinesGroupe"
+  const informationActive = currentRoute.value === "GestionnaireCantineInformations"
+  const gestionnairesActive = currentRoute.value === "GestionnaireCantineGestionnaires"
+  const pagePubliqueActive = currentRoute.value === "GestionnaireCantinePagePublique"
+  const toutesTeledeclarationsActive = currentRoute.value === "GestionnaireCantineToutesTeledeclarations"
+  const cantinesGroupeActive = currentRoute.value === "GestionnaireCantineGroupe"
 
   const informationPage = {
     text: canteenIsGroupe.value ? "Informations du groupe" : "Mes informations",
-    to: { name: "GestionnaireBilanInformations" },
+    to: { name: "GestionnaireCantineInformations" },
     active: informationActive
   }
   const gestionnairesPage = {
@@ -70,7 +71,7 @@ const menuItems = computed(() =>  {
 </script>
 
 <template>
-  <div class="layout-sidebar-bilan">
+  <div class="layout-sidebar-canteen">
     <h1>{{ canteenName }}</h1>
     <div class="ma-cantine--flex-start ma-cantine--flex-gap-1 fr-mb-4w">
       <DsfrBadge v-if="canteenBadgeGroupe" type="info" :noIcon="true" :label="canteenBadgeGroupe" />
@@ -79,24 +80,27 @@ const menuItems = computed(() =>  {
       <DsfrBadge v-if="canteenBadgeSiren" type="neutral" :label="canteenBadgeSiren" />
     </div>
     <div class="fr-grid-row fr-grid-row--top ma-cantine--sticky__container">
-      <div class="layout-sidebar-bilan__sidebar-container fr-col-12 fr-col-md-3 ma-cantine--sticky__top fr-background-default--grey">
+      <div class="layout-sidebar-canteen__sidebar-container fr-col-12 fr-col-md-3 ma-cantine--sticky__top fr-background-default--grey">
         <DsfrSideMenu :menu-items="menuItems" buttonLabel="Voir le menu"/>
       </div>
       <section class="fr-col-12 fr-col-md-9">
-        <slot :canteenInformation="canteenInformation" :canteenIsGroupe="canteenIsGroupe"></slot>
+        <slot name="title" :canteenIsGroupe="canteenIsGroupe"></slot>
+        <AppSeparator class="fr-mt-3w fr-mb-5w" />
+        <slot name="content" :canteenInformation="canteenInformation" :canteenIsGroupe="canteenIsGroupe"></slot>
       </section>
     </div>
   </div>
 </template>
 
 <style lang="scss">
-.layout-sidebar-bilan {
+.layout-sidebar-canteen {
   &__sidebar-container {
     .fr-sidemenu__title {
       display: none !important;
     }
     .fr-sidemenu__inner {
       padding-right: 0 !important;
+      box-shadow: none !important;
     }
   }
 }
